[sosreport] add per-plugin package verification lists
Add a new mechanism for plugins to declare a per-plugin list of
packages to verify when the global --verify option is enabled.

Historically this was implemented in the package manager plugin
for each distribution (e.g. rpm.py for Red Hat distros and dpkg.py
for the Debian family). This gives limited granularity and blocks
the ability to verify only a subset of packages, or to only verify
packages for a certain set of plugins.

This change adds a new optional member to Plugin: verify_packages
which when set will generate an additional command to be run by
that plugin. The plugin calls into the PackageManager via the
policy class to determine the correct package manager command to
run.

A plugin can request optional package verification by declaring a
new 'verify_packages' member (or building one during setup()):

  plugin_name = 'foo'
  packages = ('foo', 'bar', 'baz', 'qux')
  verify_packages = ('foo', 'bar')

Verification is only performed if the --verify switch is used.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>