[Plugin] automatic verify for plugins with a packages list
If a plugin defines 'packages' but not 'verify_packages', use the
package list to automatically generate a list of packages to be
verified when --verify is given.

Only the base names are used; each entry in the package list has
the '$' end-of-line match character. This prevents verification
of very large numbers of packages for plugins with package names
that are a common prefix (e.g. 'python-' and 'perl-').

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>